fix(bookmarks): share smart folders with mobile
